test: only print uffd hint if uffd actually in use
If we fail to kill firecracker because it already die, we print a hint
about the uffd handler being potentially responsible. But this only
makes sense if a uffd handler was actually spawned, so restrict printing
of the hint to this case. While we're at it, also print the uffd logs in
this case, so that the truth of the hint can easily be verified (uffd
killed firecracker iff uffd panicked).

+++ tests/framework/microvm.py
@@ -313,10 +313,14 @@ class Microvm:
             if self.screen_pid:
                 os.kill(self.screen_pid, signal.SIGKILL)
         except:
-            LOG.error(
-                "Failed to kill Firecracker Process. Did it already die (or did the UFFD handler process die and take it down)?"
-            )
-            LOG.error(self.log_data)
+            if self.uffd_handler:
+                LOG.error(
+                    "Failed to kill Firecracker Process. Did it already die (or did the UFFD handler process die and take it down)?"
+                )
+                LOG.error("UFFD Handler Logs:\n" + self.uffd_handler.log_data)
+            else:
+                LOG.error("Failed to kill Firecracker Process. Did it already die?")
+            LOG.error("Firecracker Logs:\n" + self.log_data)
             raise
 
         # if microvm was spawned then check if it gets killed
